Rename Conf classes
