+++ src/main/scala/wdlTools/syntax/Util.scala
@@ -5,7 +5,7 @@ import java.nio.file.Path
 import wdlTools.util.Util.Verbosity._
 
 object Util {
-  case class Conf(localDirectories: Seq[Path],
-                  verbosity: Verbosity = Normal,
-                  antlr4Trace: Boolean = false)
+  case class Options(localDirectories: Seq[Path],
+                     verbosity: Verbosity = Normal,
+                     antlr4Trace: Boolean = false)
 }
